requirements.txt
blinker==1.9.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
defusedxml==0.7.1
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.2.0
opencv-python==4.10.0.84
pillow==11.0.0
pyTelegramBotAPI==4.25.0
reportlab==4.2.5
requests==2.32.3
telebot==0.0.5
urllib3==2.2.3
Werkzeug==3.1.3
youtube-transcript-api==0.6.3
yt-dlp==2024.12.6
